requirements.txt
# Flask y extensiones principales
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Base de datos
SQLAlchemy==2.0.25
psycopg[binary]==3.2.10  # Para PostgreSQL (versión moderna)
pymysql==1.1.0  # Para MySQL

# Autenticación y seguridad
PyJWT==2.8.0
Werkzeug==2.3.7
bcrypt==4.0.1

# Validación y serialización
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.1  # Para tareas asíncronas
redis==4.6.0  # Para cache y Celery

# Logging y monitoreo
structlog==23.1.0
sentry-sdk[flask]==1.32.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Desarrollo
black==23.7.0  # Formateador de código
flake8==6.0.0  # Linter
isort==5.12.0  # Organizador de imports
pre-commit==3.3.3

# Documentación
flask-restx==1.1.0  # Para documentación automática de API
sphinx==7.1.2

# Backup y migración de datos
alembic==1.12.0

# Configuración y variables de entorno
python-decouple==3.8

# Manejo de fechas
python-dateutil==2.8.2

# Validación de datos
cerberus==1.3.5

# Compresión
gunicorn==21.2.0  # Servidor WSGI para producción

# Monitoreo de rendimiento
flask-profiler==1.8.1

# Rate limiting
Flask-Limiter==3.3.1

# Cache
Flask-Caching==2.0.2

# Email
Flask-Mail==0.9.1

# Archivos y uploads
Pillow==10.4.0  # Para procesamiento de imágenes

# Criptografía
cryptography==42.0.5

# Validación de esquemas JSON
jsonschema==4.19.0

# Manejo de archivos Excel
openpyxl==3.1.2
pandas==2.2.2

# Generación de reportes PDF
reportlab==4.0.4

# WebSockets (opcional para notificaciones en tiempo real)
Flask-SocketIO==5.3.5

# Health checks
flask-healthz==1.0.1

# Métricas y monitoreo
prometheus-flask-exporter==0.23.0
